fixup: move files to utils folder out of stores
